> Modificaciones en parámetros de configuración de meta4 y ptr > Obtención de datos de PTR ventas

+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/entity/TrabajoTiendaSeccionVenta.java
@@ -1,5 +1,6 @@
 package com.inditex.rrhh.icmclcwb.model.primary.entity;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -43,6 +44,6 @@ public class TrabajoTiendaSeccionVenta {
 	
 	@NotNull
 	@Column(name = "IMPORTE", nullable = false)
-	private /* BigInteger */ Double importe;
+	private /* BigInteger */ BigInteger importe;
 
 }
